HDDS-11974. Split Container Safemode Rule into Ratis & EC Container Safemode Rules (#7951)
