+++ hadoop-hdds/server-scm/src/test/java/org/apache/hadoop/hdds/scm/safemode/TestSCMSafeModeManager.java
@@ -44,6 +44,7 @@ import org.apache.hadoop.hdds.conf.OzoneConfiguration;
 import org.apache.hadoop.hdds.protocol.DatanodeDetails;
 import org.apache.hadoop.hdds.protocol.proto.HddsProtos;
 import org.apache.hadoop.hdds.protocol.proto.HddsProtos.ReplicationFactor;
+import org.apache.hadoop.hdds.protocol.proto.HddsProtos.ReplicationType;
 import org.apache.hadoop.hdds.protocol.proto.StorageContainerDatanodeProtocolProtos;
 import org.apache.hadoop.hdds.scm.HddsTestUtils;
 import org.apache.hadoop.hdds.scm.container.ContainerInfo;
@@ -130,7 +131,7 @@ public class TestSCMSafeModeManager {
       container.setNumberOfKeys(10);
     }
     ContainerManager containerManager = mock(ContainerManager.class);
-    when(containerManager.getContainers()).thenReturn(containers);
+    when(containerManager.getContainers(ReplicationType.RATIS)).thenReturn(containers);
     scmSafeModeManager = new SCMSafeModeManager(
         config, containerManager, null, null, queue,
         serviceManager, scmContext);
@@ -169,7 +170,7 @@ public class TestSCMSafeModeManager {
       container.setNumberOfKeys(10);
     }
     ContainerManager containerManager = mock(ContainerManager.class);
-    when(containerManager.getContainers()).thenReturn(containers);
+    when(containerManager.getContainers(ReplicationType.RATIS)).thenReturn(containers);
     scmSafeModeManager = new SCMSafeModeManager(
         config, containerManager, null, null, queue,
         serviceManager, scmContext);
@@ -235,7 +236,7 @@ public class TestSCMSafeModeManager {
         serviceManager,
         Clock.system(ZoneOffset.UTC));
     ContainerManager containerManager = mock(ContainerManager.class);
-    when(containerManager.getContainers()).thenReturn(containers);
+    when(containerManager.getContainers(ReplicationType.RATIS)).thenReturn(containers);
     IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
         () -> new SCMSafeModeManager(conf, containerManager,
             pipelineManager, mockNodeManager, queue, serviceManager, scmContext));
@@ -301,7 +302,7 @@ public class TestSCMSafeModeManager {
     }
 
     ContainerManager containerManager = mock(ContainerManager.class);
-    when(containerManager.getContainers()).thenReturn(containers);
+    when(containerManager.getContainers(ReplicationType.RATIS)).thenReturn(containers);
 
     scmSafeModeManager = new SCMSafeModeManager(
         conf, containerManager, pipelineManager, mockNodeManager, queue,
@@ -438,8 +439,8 @@ public class TestSCMSafeModeManager {
     conf.setBoolean(HddsConfigKeys.HDDS_SCM_SAFEMODE_ENABLED, false);
     PipelineManager pipelineManager = mock(PipelineManager.class);
     ContainerManager containerManager = mock(ContainerManager.class);
+    when(containerManager.getContainers(ReplicationType.RATIS)).thenReturn(containers);
     NodeManager nodeManager = mock(SCMNodeManager.class);
-    when(containerManager.getContainers()).thenReturn(containers);
     scmSafeModeManager = new SCMSafeModeManager(
         conf, containerManager, pipelineManager, nodeManager, queue,
         serviceManager, scmContext);
@@ -480,7 +481,7 @@ public class TestSCMSafeModeManager {
     }
 
     ContainerManager containerManager = mock(ContainerManager.class);
-    when(containerManager.getContainers()).thenReturn(containers);
+    when(containerManager.getContainers(ReplicationType.RATIS)).thenReturn(containers);
 
     scmSafeModeManager = new SCMSafeModeManager(
         config, containerManager, null, null, queue, serviceManager, scmContext);
@@ -572,16 +573,20 @@ public class TestSCMSafeModeManager {
     // the threshold will reach 100%.
     testECContainerThreshold(containers.subList(10, 20), 1.0, data);
 
-    ContainerSafeModeRule containerSafeModeRule =
-        scmSafeModeManager.getContainerSafeModeRule();
-    assertTrue(containerSafeModeRule.validate());
+    ECContainerSafeModeRule ecContainerSafeModeRule =
+        scmSafeModeManager.getECContainerSafeModeRule();
+    assertTrue(ecContainerSafeModeRule.validate());
+
+    RatisContainerSafeModeRule ratisContainerSafeModeRule =
+        scmSafeModeManager.getRatisContainerSafeModeRule();
+    assertTrue(ratisContainerSafeModeRule.validate());
   }
 
   private void testSafeModeDataNodes(int numOfDns) throws Exception {
     OzoneConfiguration conf = new OzoneConfiguration(config);
     conf.setInt(HddsConfigKeys.HDDS_SCM_SAFEMODE_MIN_DATANODE, numOfDns);
     ContainerManager containerManager = mock(ContainerManager.class);
-    when(containerManager.getContainers()).thenReturn(containers);
+    when(containerManager.getContainers(ReplicationType.RATIS)).thenReturn(containers);
     scmSafeModeManager = new SCMSafeModeManager(
         conf, containerManager, null, null, queue,
         serviceManager, scmContext);
@@ -689,7 +694,7 @@ public class TestSCMSafeModeManager {
     pipeline = pipelineManager.getPipeline(pipeline.getId());
     MockRatisPipelineProvider.markPipelineHealthy(pipeline);
     ContainerManager containerManager = mock(ContainerManager.class);
-    when(containerManager.getContainers()).thenReturn(containers);
+    when(containerManager.getContainers(ReplicationType.RATIS)).thenReturn(containers);
 
     scmSafeModeManager = new SCMSafeModeManager(
           config, containerManager, pipelineManager, nodeManager, queue,
@@ -738,7 +743,7 @@ public class TestSCMSafeModeManager {
         mockRatisProvider);
 
     ContainerManager containerManager = mock(ContainerManager.class);
-    when(containerManager.getContainers()).thenReturn(containers);
+    when(containerManager.getContainers(ReplicationType.RATIS)).thenReturn(containers);
 
     scmSafeModeManager = new SCMSafeModeManager(
         config, containerManager, pipelineManager, nodeManager, queue,
